\name{SINRELEF.LD-internal}
\alias{SINRELEF.LD-internal}
\alias{discreB}
\alias{eap_grad_obli}
\alias{eapcondef}
\alias{Nseriesk}
\alias{ordnormulti}
\alias{pgrad}
\alias{reap_grad_obli}
\alias{relicorrec}
\alias{size}
\alias{transpose}
\alias{dataconti}
\alias{LAM}
\alias{PHI}
\alias{PSI}
\alias{Ree}
\title{
SINRELEF-internal
}
\description{
This functions are for internal use only.
}
\usage{
discreB
eap_grad_obli
eapcondef
Nseriesk
ordnormulti
pgrad
reap_grad_obli
relicorrec
size
transpose
dataconti
LAM
PHI
PSI
Ree
}
\value{
No return value for the user.
}
\author{
David Navarro-Gonzalez

Urbano Lorenzo-Seva

Pere Joan Ferrando
}
\keyword{ internal }
